Combined AuthActivity into MainActivity
