+++ app/src/main/AndroidManifest.xml
@@ -13,9 +13,6 @@
         android:theme="@style/Theme.WorkWay"
         tools:targetApi="31">
         <activity
-            android:name=".AuthActivity"
-            android:exported="false" />
-        <activity
             android:name=".HomeScreen"
             android:exported="false" />
         <activity
